CARDIO-INTELLIGENCE/modules/hack-ireland-2025/little-blue-booth/scripts/requirements.txt
google-auth-oauthlib>=0.4.6
google-auth>=2.3.3
requests>=2.26.0
python-dotenv>=0.19.0 